BACKPORT: kasan: make kasan_cache_create() work with 32-bit slab cache sizes
Upstream commit be4a7988b35db9e6f95dca818d5e94785840fb58.

If SLAB doesn't support 4GB+ kmem caches (it never did), KASAN should
not do it as well.

Link: http://lkml.kernel.org/r/20180305200730.15812-20-adobriyan@gmail.com
Cc: Andrey Ryabinin <aryabinin@virtuozzo.com>
Cc: Alexander Potapenko <glider@google.com>
Cc: Dmitry Vyukov <dvyukov@google.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Bug: 128674696
(cherry picked from commit c9dcc8241bd61dc8e2d1e9fd377812c146c377c5)
(cherry picked from commit e2725de9b18276a92bfaaea46aacc1efeb4bb849)
(cherry picked from commit 7fb0925a86a0c4a50cfe7beff98003bb34ea501a)

+++ mm/slub.c
@@ -3468,7 +3468,7 @@ static void set_cpu_partial(struct kmem_cache *s)
 static int calculate_sizes(struct kmem_cache *s, int forced_order)
 {
 	unsigned long flags = s->flags;
-	size_t size = s->object_size;
+	unsigned int size = s->object_size;
 	int order;
 
 	/*
